# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SJM Social Media Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: Get feed posts
      parameters:
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found

    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like/react to a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: OK
  /posts/{id}/comments:
    get:
      operationId: listComments
      tags: [posts]
      summary: Get comments for a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"

    post:
      operationId: createComment
      tags: [posts]
      summary: Add comment to a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
  /conversations:
    get:
      operationId: listConversations
      tags: [messages]
      summary: List conversations
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"

    post:
      operationId: createConversation
      tags: [messages]
      summary: Start a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /conversations/{id}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: Get messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # NOTIFICATIONS
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: Get notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK

  # GROUPS
...
components:
  schemas:
...
    User:
      type: object
      required: [id, fullName, username, email, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        username:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        isVerified:
          type: boolean
        isAdmin:
          type: boolean
        isBanned:
          type: boolean
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        createdAt:
          type: string
...
    Post:
      type: object
      required: [id, authorId, type, createdAt]
      properties:
        id:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        type:
          type: string
          enum: [text, image, video, poll, reel]
        content:
          type: ["string", "null"]
        mediaUrls:
          type: array
          items:
            type: string
        hashtags:
          type: array
          items:
            type: string
        likesCount:
          type: integer
        commentsCount:
          type: integer
        sharesCount:
          type: integer
        isLiked:
          type: boolean
        userReaction:
          type: ["string", "null"]
        audience:
          type: string
          enum: [public, friends, private]
        createdAt:
          type: string
    PostInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [text, image, video, poll]
        content:
          type: ["string", "null"]
        mediaUrls:
          type: array
          items:
            type: string
        hashtags:
          type: array
          items:
            type: string
        audience:
          type: string
          enum: [public, friends, private]
    ReactionInput:
      type: object
      required: [reaction]
      properties:
        reaction:
          type: string
          enum: [like, love, haha, wow, sad, angry]
    Comment:
      type: object
      required: [id, postId, authorId, content, createdAt]
      properties:
        id:
          type: integer
        postId:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        content:
          type: string
        parentId:
          type: ["integer", "null"]
        likesCount:
          type: integer
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        parentId:
          type: ["integer", "null"]

    # STORY
...
    Conversation:
      type: object
      required: [id, createdAt]
      properties:
        id:
          type: integer
        participants:
          type: array
          items:
            $ref: "#/components/schemas/User"
        isGroup:
          type: boolean
        groupName:
          type: ["string", "null"]
        groupAvatarUrl:
          type: ["string", "null"]
        lastMessage:
          $ref: "#/components/schemas/Message"
        unreadCount:
          type: integer
        createdAt:
          type: string
    ConversationInput:
      type: object
      required: [participantIds]
      properties:
        participantIds:
          type: array
          items:
            type: integer
        isGroup:
          type: boolean
        groupName:
          type: ["string", "null"]
    Message:
      type: object
      required: [id, conversationId, senderId, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        sender:
          $ref: "#/components/schemas/User"
        content:
          type: string
        type:
          type: string
          enum: [text, image, video, audio, file, emoji]
        mediaUrl:
          type: ["string", "null"]
        isRead:
          type: boolean
        isDeleted:
          type: boolean
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        type:
          type: string
          enum: [text, image, video, audio, file, emoji]
        mediaUrl:
          type: ["string", "null"]

    # NOTIFICATIONS
    Notification:
      type: object
      required: [id, userId, type, message, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [like, comment, follow, message, mention, group_invite, story_reaction, reel_like]
        message:
          type: string
        actorId:
          type: ["integer", "null"]
        actor:
          $ref: "#/components/schemas/User"
        entityId:
          type: ["integer", "null"]
        entityType:
          type: ["string", "null"]
        isRead:
          type: boolean
        createdAt:
          type: string

    # GROUPS
...